Fixed an annoying bug where when adding a new resident the medlist wasn't getting cleared
