--- src/components/Medicine/MedicinePage.js
@@ -1,279 +1,0 @@
-import React, {useGlobal, useState, useEffect} from 'reactn';
-import Button from 'react-bootstrap/Button';
-import Row from 'react-bootstrap/Row';
-import Col from 'react-bootstrap/Col';
-import Form from 'react-bootstrap/Form';
-import TabContent from "../../styles/tab_content.css";
-import RefreshMedicineList from "../../providers/RefreshMedicineList";
-import MedicineEdit from "../Medicine/MedicineEdit";
-import ConfirmationDialog from "../Dialog/ConfirmationDialog";
-import DrugLogGrid from "../DrugLog/DrugLogGrid";
-import DrugLogEdit from "../DrugLog/DrugLogEdit";
-import RefreshMedicineLog from "../../providers/RefreshMedicineLog";
-import MedicineListGroup from "./MedicineListGroup";
-import TooltipButton from "../Buttons/TooltipButton";
-import {calculateLastTaken} from "../../utility/common";
-import {newDrugInfo} from "../../utility/InitialState";
-
-/**
- * MedicinePage
- * This is where most of the logic and data entry will be done so the page is a little busy
- * Features of this page are:
- *  - Manually Add Medicine
- *  - Add Drug to Medicine Log
- *  - Edit / Delete Existing Medicine
- *  - Drug History Grid
- *
- * @returns {*}
- */
-const MedicinePage = (props) => {
-    const [ showMedicineEdit, setShowMedicineEdit ] = useState(false);
-    const [ showDrugLog, setShowDrugLog ] = useState(false);
-    const [ drugInfo, setDrugInfo ] = useState(null);
-    const [ drugLogInfo, setDrugLogInfo ] = useState(null);
-    const [ showDeleteDrugLogRecord, setShowDeleteDrugLogRecord ] = useState(false);
-    const [ lastTaken, setLastTaken ] = useState(false);
-    const [ activeDrug, setActiveDrug ] = useState(null);
-
-    const [ medicineList, setMedicineList ] = useGlobal('medicineList');
-    const [ drugLogList, setDrugLogList ] = useGlobal('drugLogList');
-    const [ activeResident ] = useGlobal('activeResident');
-    const [ providers ] = useGlobal('providers');
-
-    const medHistoryProvider = providers.medHistoryProvider;
-    const medicineProvider = providers.medicineProvider;
-
-    useEffect(()=> {
-        if (medicineList && medicineList.length > 0) {
-            setActiveDrug(medicineList[0]);
-        }
-    }, [medicineList]);
-
-    // Calculate how many hours it has been since the activeDrug was taken and set showLastTakenWarning value
-    useEffect(() => {
-        if (activeDrug && activeDrug.Id && drugLogList) {
-            setLastTaken(calculateLastTaken(activeDrug.Id, drugLogList));
-        } else {
-            setLastTaken(null);
-        }
-    }, [activeDrug, drugLogList]);
-
-    /**
-     * Fires when medicine is added or edited.
-     *
-     * @param {Event} e
-     * @param {boolean} isAdd
-     */
-    function addEditDrug(e, isAdd)
-    {
-        e.preventDefault();
-        if (isAdd) {
-            const drugInfo = {...newDrugInfo};
-            drugInfo.OTC = false;
-            drugInfo.ResidentId = activeResident.Id;
-            setDrugInfo(drugInfo);
-        } else {
-            setDrugInfo({...activeDrug});
-        }
-        setShowMedicineEdit(true);
-    }
-
-    /**
-     * Fires when MedicineEdit closes.
-     *
-     * @param {object | null} drugInfo
-     */
-    function handleMedicineEditModalClose(drugInfo)
-    {
-        if (drugInfo) {
-            const drugData = {...drugInfo};
-
-            if (!drugData.Id) {
-                drugData.Id = null;
-            }
-
-            if (drugData.Notes === '') {
-                drugData.Notes = null;
-            }
-
-            medicineProvider.post(drugData)
-            .then((drugRecord) => {
-                RefreshMedicineList(medicineProvider, drugData.ResidentId)
-                .then((drugList) => {
-                    setMedicineList(drugList);
-                    setDrugInfo(drugRecord);
-                    setActiveDrug(drugRecord);
-                    setLastTaken(false);
-                    RefreshMedicineLog(medHistoryProvider, drugData.ResidentId)
-                        .then((updatedDrugLog) => setDrugLogList(updatedDrugLog));
-                })
-                .catch((err) => {
-                    props.onError(err);
-                });
-            });
-        }
-        setShowMedicineEdit(false);
-    }
-
-    /**
-     * Fires when the user has confirmed the deletion of a drug log record.
-     *
-     * @param {object} drugLogInfo
-     */
-    function deleteDrugLogRecord(drugLogInfo)
-    {
-        medHistoryProvider.delete(drugLogInfo.Id)
-        .then((response) => {
-            RefreshMedicineLog(medHistoryProvider, activeDrug.ResidentId)
-                .then((data) => setDrugLogList(data));
-        });
-        setShowDeleteDrugLogRecord(false);
-    }
-
-    /**
-     * Fires when user clicks on +Log or the drug log edit button
-     *
-     * @param {Event} e
-     * @param {object} drugLogInfo
-     */
-    function addEditDrugLog(e, drugLogInfo)
-    {
-        e.preventDefault();
-
-        // If drugLogInfo is not populated then this is an add operation.
-        if (!drugLogInfo) {
-            drugLogInfo = {
-                Id: null,
-                ResidentId: activeResident.Id,
-                MedicineId: activeDrug.Id,
-                Notes: ""
-            }
-        }
-
-        setDrugLogInfo(drugLogInfo);
-        setShowDrugLog(true);
-    }
-
-    /**
-     * Fires when the drug log edit modal closes
-     *
-     * @param {object | null} drugLogInfo
-     */
-    function handleDrugLogEditClose(drugLogInfo) {
-        if (drugLogInfo) {
-            medHistoryProvider.post(drugLogInfo)
-                .then((response) => {
-                    RefreshMedicineLog(medHistoryProvider, activeDrug.ResidentId)
-                        .then((data) => setDrugLogList(data));
-                })
-                .catch((err) => {
-                    props.onError(err);
-                });
-        }
-        setShowDrugLog(false);
-    }
-
-    return (
-        <>
-            <Form className={TabContent}>
-                {activeResident && activeResident.Id &&
-                <Row controlId="medicine-buttons">
-                    <Col sm="4">
-                        <>
-                            <TooltipButton
-                                tooltip="Manually Add New Medicine"
-                                size="sm"
-                                variant="info"
-                                onClick={(e) => addEditDrug(e, true)}
-                            >
-                                + Medicine
-                            </TooltipButton>
-
-                            {activeDrug &&
-                            <Button
-                                className="ml-2"
-                                size="sm"
-                                variant="info"
-                                onClick={(e) => addEditDrug(e, false)}
-                            >
-                                Edit <b>{activeDrug.Drug}</b>
-                            </Button>
-                            }
-                        </>
-                    </Col>
-
-                    {activeDrug &&
-                    <Col sm="8">
-                        <span style={{textAlign: "center"}}> <h2>{activeDrug.Drug} History</h2> </span>
-                    </Col>
-                    }
-                </Row>
-                }
-
-                {activeResident && activeResident.Id && activeDrug &&
-                <Row>
-                    <Col sm="4">
-                        {medicineList && activeDrug &&
-                            <MedicineListGroup
-                                lastTaken={lastTaken}
-                                medicineList={medicineList}
-                                activeDrug={activeDrug}
-                                drugChanged={(drug) => setActiveDrug(drug)}
-                                addDrugLog={(e) => addEditDrugLog(e)}
-                            />
-                        }
-                    </Col>
-
-                    <Col sm="8">
-                        <DrugLogGrid
-                            showDrugColumn={false}
-                            drugLog={drugLogList}
-                            drugId={activeDrug && activeDrug.Id}
-                            onEdit={(e, r) => addEditDrugLog(e, r)}
-                            onDelete={(e, r) => setShowDeleteDrugLogRecord(r)}
-                        />
-                    </Col>
-                </Row>
-                }
-            </Form>
-
-            {/* MedicineEdit Modal*/}
-            <MedicineEdit
-                show={showMedicineEdit}
-                onHide={() => setShowMedicineEdit(!showMedicineEdit)}
-                onClose={(r) => handleMedicineEditModalClose(r)}
-                drugInfo={drugInfo}
-            />
-
-            <DrugLogEdit
-                show={showDrugLog}
-                drugLogInfo={drugLogInfo}
-                onHide={() => setShowDrugLog(!showDrugLog)}
-                onClose={(drugLogRecord) => handleDrugLogEditClose(drugLogRecord)}
-            />
-
-            <ConfirmationDialog
-                title="Delete Log Record"
-                body={
-                    <>
-                        <p>{showDeleteDrugLogRecord.Created}</p>
-                        <b style={{color: "red"}}>
-                            Are you sure?
-                        </b>
-                    </>
-                }
-                show={showDeleteDrugLogRecord}
-                onAnswer={(a) => {
-                    if (a) {
-                        deleteDrugLogRecord(showDeleteDrugLogRecord);
-                    } else {
-                        setShowDeleteDrugLogRecord(false);
-                    }
-                }}
-                onHide={() => setShowDeleteDrugLogRecord(false)}
-            />
-        </>
-    );
-}
-
-export default MedicinePage;
