+++ src/components/Modals/Dialog/InformationDialog.js
